Fix ambiguous pgr_dijkstra overload in /api/directions
node-postgres sends query params as untyped, which left Postgres
unable to pick between pgr_dijkstra's scalar/array vid overloads
(REQUEST failed with "function pgr_dijkstra(...) is not unique").
Casting the vid params to bigint resolves it.

+++ backend/server.js
@@ -315,7 +315,7 @@ app.get('/api/directions', async (req, res) => {
                     SUM(d.cost) AS total_cost_m
              FROM pgr_dijkstra(
                  'SELECT way_id AS id, source, target, cost, reverse_cost FROM osm_roads',
-                 $1, $2
+                 $1::bigint, $2::bigint
              ) d
              JOIN osm_roads r ON r.way_id = d.edge
              WHERE d.edge != -1`,
